requirements.txt
Django==5.2.4
whitenoise==6.11.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
dj-database-url==3.0.1
Pillow==11.3.0
reportlab==4.4.4
python-dateutil==2.9.0.post0
django-crontab==0.7.1
